% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_gene_annotation.R
\name{annotate_via_msigdbr}
\alias{annotate_via_msigdbr}
\title{Annotate genes using msigdbr mapping table}
\usage{
annotate_via_msigdbr(ensembl_ids)
}
\arguments{
\item{ensembl_ids}{Character vector of clean Ensembl IDs}
}
\value{
Data frame with ensembl_gene, gene_symbol, entrez_gene
}
\description{
Annotate genes using msigdbr mapping table
}
\keyword{internal}
