% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_data.R
\name{classify_regimen}
\alias{classify_regimen}
\title{Classify regimen by drug class}
\usage{
classify_regimen(regimen_std)
}
\arguments{
\item{regimen_std}{Character vector of standardized regimen names}
}
\value{
Character vector of regimen classes
}
\description{
Classify regimen by drug class
}
\keyword{internal}
