% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMMpass-package.R
\docType{package}
\name{coMMpass-package}
\alias{coMMpass}
\alias{coMMpass-package}
\title{coMMpass: MMRF CoMMpass Data Analysis Pipeline}
\description{
A reproducible data acquisition and analysis pipeline for the MMRF CoMMpass study using Targets and Nix.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://JohnGavin.github.io/coMMpass-analysis/}
  \item \url{https://github.com/JohnGavin/coMMpass-analysis}
  \item Report bugs at \url{https://github.com/JohnGavin/coMMpass-analysis/issues}
}

}
\author{
\strong{Maintainer}: John Gavin \email{john.gavin@example.com}

}
\keyword{internal}
