% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_acquisition.R
\name{download_aws_data}
\alias{download_aws_data}
\title{Download data from AWS S3 open access bucket}
\usage{
download_aws_data(
  bucket_name = "gdc-mmrf-commpass-phs000748-2-open",
  prefix = NULL,
  data_dir = "data/raw/aws",
  region = "us-east-1"
)
}
\arguments{
\item{bucket_name}{S3 bucket name}

\item{prefix}{File prefix to filter}

\item{data_dir}{Local directory for downloads}

\item{region}{AWS region}
}
\value{
Directory path as character string
}
\description{
Download data from AWS S3 open access bucket
}
