% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract counts matrix from SummarizedExperiment}
\usage{
extract_counts(se_data)
}
\arguments{
\item{se_data}{SummarizedExperiment object}
}
\value{
Integer matrix of counts, or NULL
}
\description{
Extract counts matrix from SummarizedExperiment
}
\keyword{internal}
