% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_quality_control.R
\name{filter_low_quality}
\alias{filter_low_quality}
\title{Filter low-quality samples and genes}
\usage{
filter_low_quality(se_data, min_counts = 10, min_samples = 3)
}
\description{
Filter low-quality samples and genes
}
