% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{find_consensus_genes}
\alias{find_consensus_genes}
\title{Find consensus DE genes across methods}
\usage{
find_consensus_genes(de_results_list, padj_threshold = 0.05, lfc_threshold = 1)
}
\arguments{
\item{de_results_list}{List of DE result objects from different methods}

\item{padj_threshold}{Adjusted p-value threshold (default: 0.05)}

\item{lfc_threshold}{Log2 fold change threshold (default: 1)}
}
\value{
List with consensus gene information
}
\description{
Find consensus DE genes across methods
}
