% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_de_visualization.R
\name{find_expr_col}
\alias{find_expr_col}
\title{Find the mean expression column name}
\usage{
find_expr_col(df)
}
\arguments{
\item{df}{Data frame with DE results}
}
\value{
Column name string
}
\description{
Find the mean expression column name
}
\keyword{internal}
