% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_de_visualization.R
\name{find_lfc_col}
\alias{find_lfc_col}
\title{Find the log fold-change column name}
\usage{
find_lfc_col(df)
}
\arguments{
\item{df}{Data frame with DE results}
}
\value{
Column name string
}
\description{
Find the log fold-change column name
}
\keyword{internal}
