% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_quality_control.R
\name{normalize_rnaseq}
\alias{normalize_rnaseq}
\title{Normalize RNA-seq data}
\usage{
normalize_rnaseq(se_data, method = "TMM")
}
\arguments{
\item{se_data}{A SummarizedExperiment object with a "counts" assay}

\item{method}{Normalization method (default: "TMM")}
}
\value{
SummarizedExperiment with added "logCPM" assay
}
\description{
Normalize RNA-seq data
}
