% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_gene_correlations.R
\name{plot_gene_correlation}
\alias{plot_gene_correlation}
\title{Scatter plot of gene-gene correlation}
\usage{
plot_gene_correlation(cor_result, title = NULL)
}
\arguments{
\item{cor_result}{List from [correlate_genes()]}

\item{title}{Plot title (default auto-generated)}
}
\value{
A ggplot object
}
\description{
Creates a scatter plot with regression line, correlation coefficient,
and p-value annotation from a [correlate_genes()] result.
}
\seealso{
Other gene-correlation: 
\code{\link{correlate_genes}()},
\code{\link{correlate_genes_batch}()}
}
\concept{differential-expression}
\concept{gene-correlation}
