% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{prepare_coldata}
\alias{prepare_coldata}
\title{Prepare colData for DESeq2}
\usage{
prepare_coldata(se_data, clinical_data, design_formula)
}
\arguments{
\item{se_data}{SummarizedExperiment object}

\item{clinical_data}{Clinical data frame}

\item{design_formula}{Design formula}
}
\value{
Data frame suitable for DESeq2 colData
}
\description{
Prepare colData for DESeq2
}
\keyword{internal}
