% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{query_commpass_rna}
\alias{query_commpass_rna}
\title{Query GDC for CoMMpass RNA-seq Metadata}
\usage{
query_commpass_rna()
}
\value{
A GDCquery object containing metadata for RNA-seq samples
}
\description{
Queries the Genomic Data Commons (GDC) API for Multiple Myeloma Research
Foundation (MMRF) CoMMpass study RNA-seq data. This function returns a
query object that can be used with other TCGAbiolinks functions to download
and prepare the data.
}
\examples{
\dontrun{
# Query for RNA-seq data
query <- query_commpass_rna()

# Use the query to download data
# GDCdownload(query)
# data <- GDCprepare(query)
}
}
\seealso{
Other data-acquisition: 
\code{\link{acquire_commpass_data}()},
\code{\link{download_clinical_data}()},
\code{\link{download_gdc_rnaseq}()},
\code{\link{download_s3_subset}()},
\code{\link{get_commpass_clinical}()},
\code{\link{list_s3_commpass}()}
}
\concept{data-acquisition}
