% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_pathway_analysis.R
\name{run_gsea}
\alias{run_gsea}
\title{Run Gene Set Enrichment Analysis}
\usage{
run_gsea(se_data, clinical_data)
}
\arguments{
\item{se_data}{A SummarizedExperiment object}

\item{clinical_data}{Clinical data frame}
}
\value{
List with GSEA results
}
\description{
Run Gene Set Enrichment Analysis
}
