% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_timestamped}
\alias{save_timestamped}
\title{Save results with timestamp}
\usage{
save_timestamped(object, base_name, dir = "results")
}
\arguments{
\item{object}{R object to save}

\item{base_name}{Base name for the output file}

\item{dir}{Directory to save to (default: "results")}
}
\description{
Save results with timestamp
}
