% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_logging}
\alias{setup_logging}
\title{Setup logging}
\usage{
setup_logging(log_file = NULL)
}
\arguments{
\item{log_file}{Optional path to a log file. If NULL, logs to console.}
}
\description{
Setup logging
}
