% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_data.R
\name{standardize_response}
\alias{standardize_response}
\title{Standardize best response values}
\usage{
standardize_response(response)
}
\arguments{
\item{response}{Character vector of raw response values}
}
\value{
Ordered factor of standardized responses
}
\description{
Maps raw response strings to an ordered factor: sCR > CR > VGPR > PR >
MR > SD > PD.
}
\keyword{internal}
