% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_data}
\alias{summarize_data}
\title{Generate summary statistics}
\usage{
summarize_data(se_data)
}
\arguments{
\item{se_data}{A SummarizedExperiment object with a "counts" assay}
}
\description{
Generate summary statistics
}
