% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_gene_annotation.R
\name{annotate_de_results}
\alias{annotate_de_results}
\title{Add gene symbol column to DE results table}
\usage{
annotate_de_results(de_results, annotation)
}
\arguments{
\item{de_results}{Data frame with Ensembl IDs as rownames}

\item{annotation}{Data frame from annotate_genes()}
}
\value{
DE results with gene_symbol column prepended
}
\description{
Add gene symbol column to DE results table
}
\seealso{
Other pathway: 
\code{\link{annotate_genes}()},
\code{\link{plot_enrichment_barplot}()},
\code{\link{plot_enrichment_dotplot}()},
\code{\link{plot_gsea_running_score}()},
\code{\link{run_gsea}()},
\code{\link{run_ora}()},
\code{\link{run_pathway_analysis}()}
}
\concept{pathway}
