% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_quality_control.R
\name{calculate_qc_metrics}
\alias{calculate_qc_metrics}
\title{Calculate QC metrics for RNA-seq data}
\usage{
calculate_qc_metrics(se_data)
}
\arguments{
\item{se_data}{A SummarizedExperiment object with a "counts" assay}
}
\value{
A data frame with QC metrics per sample
}
\description{
Calculate QC metrics for RNA-seq data
}
