% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_clinical_data}
\alias{clean_clinical_data}
\title{Clean Clinical Data}
\usage{
clean_clinical_data(clinical_raw)
}
\arguments{
\item{clinical_raw}{Raw clinical data frame}
}
\value{
Cleaned clinical data frame
}
\description{
Standardizes clinical data column names and formats
}
\seealso{
Other data-cleaning: 
\code{\link{clean_expression_data}()},
\code{\link{clean_treatment_data}()},
\code{\link{integrate_clinical_expression}()},
\code{\link{summarize_treatment}()}
}
\concept{data-cleaning}
