% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_expression_data}
\alias{clean_expression_data}
\title{Clean Expression Data}
\usage{
clean_expression_data(expr_raw)
}
\arguments{
\item{expr_raw}{Raw expression data (matrix or data frame)}
}
\value{
Cleaned expression matrix with gene names as rownames
}
\description{
Standardizes expression data format and adds metadata
}
\seealso{
Other data-cleaning: 
\code{\link{clean_clinical_data}()},
\code{\link{clean_treatment_data}()},
\code{\link{integrate_clinical_expression}()},
\code{\link{summarize_treatment}()}
}
\concept{data-cleaning}
