% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{download_s3_subset}
\alias{download_s3_subset}
\title{Download a sample of RNA-seq files from S3}
\usage{
download_s3_subset(s3_paths, dest_dir = "data/raw/rna_seq", n = 3)
}
\arguments{
\item{s3_paths}{Character vector of S3 keys (files)}

\item{dest_dir}{Destination directory}

\item{n}{Number of files to download}
}
\description{
Download a sample of RNA-seq files from S3
}
