% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{find_consensus_genes}
\alias{find_consensus_genes}
\title{Find consensus DE genes across methods}
\usage{
find_consensus_genes(de_results_list, padj_threshold = 0.05, lfc_threshold = 1)
}
\description{
Find consensus DE genes across methods
}
