% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_with_commas}
\alias{format_with_commas}
\title{Format Number with Thousands Separator}
\usage{
format_with_commas(x)
}
\arguments{
\item{x}{Numeric value or vector}
}
\value{
Character vector with formatted numbers
}
\description{
Adds commas as thousands separators to large numbers
}
\examples{
format_with_commas(1234567)
# Returns: "1,234,567"
}
\seealso{
Other utilities: 
\code{\link{create_summary_table}()},
\code{\link{example_data}()},
\code{\link{format_file_size}()},
\code{\link{strip_plotly}()}
}
\concept{utilities}
