% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_causal_dags.R
\name{get_adjustment_sets}
\alias{get_adjustment_sets}
\title{Get adjustment sets for a given analysis}
\usage{
get_adjustment_sets(
  exposure = "cytogenetic_risk",
  outcome = "overall_survival",
  dag = commpass_dag()
)
}
\arguments{
\item{exposure}{Character. The exposure variable name in the DAG.}

\item{outcome}{Character. The outcome variable name in the DAG.}

\item{dag}{A `dagitty` DAG object. Defaults to [commpass_dag()].}
}
\value{
A list of character vectors, each a minimal sufficient adjustment set.
}
\description{
Uses the DAG to identify the minimal sufficient adjustment set for
estimating the causal effect of `exposure` on `outcome`.
}
