% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_commpass_clinical}
\alias{get_commpass_clinical}
\title{Query GDC for CoMMpass Clinical Data}
\usage{
get_commpass_clinical()
}
\value{
A data frame of clinical data
}
\description{
Query GDC for CoMMpass Clinical Data
}
