% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_counts_assay}
\alias{get_counts_assay}
\title{Get counts assay from SummarizedExperiment}
\usage{
get_counts_assay(se)
}
\arguments{
\item{se}{SummarizedExperiment object}
}
\value{
Counts matrix
}
\description{
GDC STAR-Counts uses 'unstranded', but fallback to 'counts' if present
}
