% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{integrate_clinical_expression}
\alias{integrate_clinical_expression}
\title{Create Integrated Dataset}
\usage{
integrate_clinical_expression(clinical_clean, expr_clean)
}
\arguments{
\item{clinical_clean}{Cleaned clinical data}

\item{expr_clean}{Cleaned expression data}
}
\value{
List with matched clinical and expression data
}
\description{
Combines clinical and expression data with consistent sample IDs
}
