% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{list_s3_commpass}
\alias{list_s3_commpass}
\title{List AWS S3 CoMMpass Bucket}
\usage{
list_s3_commpass(prefix = "")
}
\arguments{
\item{prefix}{Optional prefix to filter}
}
\value{
Character vector of file paths
}
\description{
List AWS S3 CoMMpass Bucket
}
