% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_causal_dags.R
\name{plot_dag}
\alias{plot_dag}
\title{Plot the CoMMpass causal DAG}
\usage{
plot_dag(dag = NULL, title = "CoMMpass Causal DAG")
}
\arguments{
\item{dag}{A `dagitty` DAG object. Defaults to [commpass_dag()].}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Renders the DAG using ggdag with sensible defaults.
}
