% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_gene_annotation.R
\name{plot_enrichment_barplot}
\alias{plot_enrichment_barplot}
\title{Bar plot of enrichment results}
\usage{
plot_enrichment_barplot(enrich_result, n = 15L, title = "Enrichment Bar Plot")
}
\arguments{
\item{enrich_result}{Data frame with pathway enrichment results}

\item{n}{Number of top pathways to show (default 15)}

\item{title}{Plot title}
}
\value{
A ggplot2 object
}
\description{
Creates a horizontal bar plot of the top N enriched pathways, colored
by direction (GSEA) or significance (ORA).
}
\seealso{
Other pathway: 
\code{\link{annotate_de_results}()},
\code{\link{annotate_genes}()},
\code{\link{plot_enrichment_dotplot}()},
\code{\link{plot_gsea_running_score}()},
\code{\link{run_gsea}()},
\code{\link{run_ora}()},
\code{\link{run_pathway_analysis}()}
}
\concept{pathway}
