% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_survival_analysis.R
\name{prepare_survival_data}
\alias{prepare_survival_data}
\title{Prepare survival data}
\usage{
prepare_survival_data(clinical_data, se_data)
}
\arguments{
\item{clinical_data}{Clinical data frame}

\item{se_data}{A SummarizedExperiment object (optional)}
}
\value{
Data frame with survival time, status, and covariates
}
\description{
Prepare survival data
}
