% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{query_commpass_rna}
\alias{query_commpass_rna}
\title{Query GDC for CoMMpass RNA-seq Metadata}
\usage{
query_commpass_rna()
}
\value{
A GDC query object
}
\description{
Query GDC for CoMMpass RNA-seq Metadata
}
