% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run DESeq2 differential expression analysis}
\usage{
run_deseq2(se_data, clinical_data, design_formula = ~condition)
}
\description{
Run DESeq2 differential expression analysis
}
