% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{run_edger}
\alias{run_edger}
\title{Run edgeR differential expression analysis}
\usage{
run_edger(se_data, clinical_data, design_formula = ~condition)
}
\description{
Run edgeR differential expression analysis
}
