% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_differential_expression.R
\name{run_limma}
\alias{run_limma}
\title{Run limma differential expression analysis}
\usage{
run_limma(se_data, clinical_data, design_formula = ~condition)
}
\description{
Run limma differential expression analysis
}
