% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_pathway_analysis.R
\name{run_pathway_analysis}
\alias{run_pathway_analysis}
\title{Run pathway enrichment analysis}
\usage{
run_pathway_analysis(de_genes, method = "clusterProfiler")
}
\description{
Run pathway enrichment analysis
}
